# CUDA: Optimize `rms_norm_f32` kernel and its fused variants, giving 1-6% perf E2E