--- a/ggml/src/ggml-cuda/common.cuh
+++ b/ggml/src/ggml-cuda/common.cuh
@@ -569,25 +569,33 @@ static __device__ __forceinline__ float ggml_cuda_e8m0_to_fp32(uint8_t x) {
 // and a shift:
 //
 // n/d = (mulhi(n, mp) + n) >> L;
-static void init_fastdiv_values(uint32_t d, uint32_t & mp, uint32_t & L) {
+static const uint2 init_fastdiv_values(uint32_t d) {
     // compute L = ceil(log2(d));
-    L = 0;
+    uint32_t L = 0;
     while (L < 32 && (uint32_t{ 1 } << L) < d) {
         L++;
     }
 
-    mp = (uint32_t) ((uint64_t{ 1 } << 32) * ((uint64_t{ 1 } << L) - d) / d + 1);
+    uint32_t mp = (uint32_t) ((uint64_t{ 1 } << 32) * ((uint64_t{ 1 } << L) - d) / d + 1);
+    return make_uint2(mp, L);
 }
 
-static __device__ __forceinline__ uint32_t fastdiv(uint32_t n, uint32_t mp, uint32_t L) {
+static __device__ __forceinline__ uint32_t fastdiv(uint32_t n, const uint2 div_consts) {
     // Compute high 32 bits of n * mp
-    const uint32_t hi = __umulhi(n, mp);
+    const uint32_t hi = __umulhi(n, div_consts.x);
     // Apply the formula
-    return (hi + n) >> L;
+    return (hi + n) >> div_consts.y;
 }
 
-static __device__ __forceinline__ uint32_t fastmodulo(uint32_t n, uint32_t divisor, int mp, uint32_t L) {
-    return n - fastdiv(n, mp, L) * divisor;
+static const uint3 init_fastmodulo_values(uint32_t d) {
+    // uint3 contains <mp, L, divisor> in <x, y, z>
+    const uint2 fastdiv = init_fastdiv_values(d);
+    return make_uint3(fastdiv.x, fastdiv.y, d);
+}
+
+static __device__ __forceinline__ uint32_t fastmodulo(uint32_t n, const uint3 div_consts_divisor) {
+    // expects div_consts_divisor to contain <mp, L, divisor> in <x, y, z>
+    return n - fastdiv(n, make_uint2(div_consts_divisor.x, div_consts_divisor.y)) * div_consts_divisor.z;
 }
 
 typedef void (*dequantize_kernel_t)(const void * vx, const int64_t ib, const int iqs, float2 & v);