This PR optimizes `rms_norm_f32` kernel and its fused variants by the following 2 changes:

1. Addition of [fast integer division](https://en.wikipedia.org/wiki/Division_algorithm#Division_by_a_constant), and using it within the modulo operator. Integer division by run-time constants can be implemented via a single multiplication + bit-shift, which is significantly faster than the naive implementation. By pre-computing values on host/CPU side, we thus gain a lot of perf, especially when the same division is performed multiple times within a kernel. This technique is already used in our [vulkan](https://github.com/ggml-org/llama.cpp/blob/77dee9de97be75b7143a213bc48893e0c0b29af7/ggml/src/ggml-vulkan/ggml-vulkan.cpp#L794-L809) backend, and can be applied to other cuda kernels which use the same modulo operation multiple times in future.
2. Adding support for arbitrary values of `blockDim.x` in the interval `[1, 1024]`. This allows us to schedule 2 full warps on a SM for small vector-lengths, better hiding latencies there.

---

Together, this leads to 1-6 % perf gains across the following benched models:

master (b66df9d9)
  Device 0: NVIDIA RTX PRO 6000 Blackwell Max-Q Workstation Edition, compute capability 12.0, VMM: yes
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma3n E2B Q8_0               |   4.45 GiB |     4.46 B | CUDA       |  99 |    pp100 @ d100 |      3545.35 ± 41.07 |
| gemma3n E2B Q8_0               |   4.45 GiB |     4.46 B | CUDA       |  99 |    tg100 @ d100 |        165.46 ± 0.69 |
| gemma3n E4B Q8_0               |   6.84 GiB |     6.87 B | CUDA       |  99 |    pp100 @ d100 |      2587.48 ± 76.47 |
| gemma3n E4B Q8_0               |   6.84 GiB |     6.87 B | CUDA       |  99 |    tg100 @ d100 |        124.24 ± 0.48 |
| gemma3 12B Q8_0                |  11.64 GiB |    11.77 B | CUDA       |  99 |    pp100 @ d100 |      2899.78 ± 18.41 |
| gemma3 12B Q8_0                |  11.64 GiB |    11.77 B | CUDA       |  99 |    tg100 @ d100 |         86.09 ± 0.13 |
| llama 3B Q4_K - Medium         |   1.87 GiB |     3.21 B | CUDA       |  99 |    pp100 @ d100 |     6281.45 ± 174.32 |
| llama 3B Q4_K - Medium         |   1.87 GiB |     3.21 B | CUDA       |  99 |    tg100 @ d100 |        327.53 ± 1.58 |
| qwen3 4B Q4_K - Medium         |   2.44 GiB |     4.02 B | CUDA       |  99 |    pp100 @ d100 |      5155.90 ± 51.79 |
| qwen3 4B Q4_K - Medium         |   2.44 GiB |     4.02 B | CUDA       |  99 |    tg100 @ d100 |        242.80 ± 2.91 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |    pp100 @ d100 |      3293.58 ± 64.60 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |    tg100 @ d100 |        242.56 ± 0.72 |
| qwen3moe 30B.A3B Q3_K - Small  |  12.37 GiB |    30.53 B | CUDA       |  99 |    pp100 @ d100 |      1862.04 ± 45.24 |
| qwen3moe 30B.A3B Q3_K - Small  |  12.37 GiB |    30.53 B | CUDA       |  99 |    tg100 @ d100 |        165.02 ± 0.84 |

This PR:
  Device 0: NVIDIA RTX PRO 6000 Blackwell Max-Q Workstation Edition, compute capability 12.0, VMM: yes
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gemma3n E2B Q8_0               |   4.45 GiB |     4.46 B | CUDA       |  99 |    pp100 @ d100 |      3649.65 ± 30.64 |
| gemma3n E2B Q8_0               |   4.45 GiB |     4.46 B | CUDA       |  99 |    tg100 @ d100 |        175.12 ± 1.08 |
| gemma3n E4B Q8_0               |   6.84 GiB |     6.87 B | CUDA       |  99 |    pp100 @ d100 |     2871.36 ± 137.86 |
| gemma3n E4B Q8_0               |   6.84 GiB |     6.87 B | CUDA       |  99 |    tg100 @ d100 |        130.39 ± 0.33 |
| gemma3 12B Q8_0                |  11.64 GiB |    11.77 B | CUDA       |  99 |    pp100 @ d100 |      3013.22 ± 15.47 |
| gemma3 12B Q8_0                |  11.64 GiB |    11.77 B | CUDA       |  99 |    tg100 @ d100 |         89.35 ± 0.12 |
| llama 3B Q4_K - Medium         |   1.87 GiB |     3.21 B | CUDA       |  99 |    pp100 @ d100 |      6458.34 ± 26.40 |
| llama 3B Q4_K - Medium         |   1.87 GiB |     3.21 B | CUDA       |  99 |    tg100 @ d100 |       330.64 ± 11.34 |
| qwen3 4B Q4_K - Medium         |   2.44 GiB |     4.02 B | CUDA       |  99 |    pp100 @ d100 |      5200.88 ± 38.18 |
| qwen3 4B Q4_K - Medium         |   2.44 GiB |     4.02 B | CUDA       |  99 |    tg100 @ d100 |        253.41 ± 0.61 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |    pp100 @ d100 |     3244.66 ± 102.53 |
| gpt-oss 20B MXFP4 MoE          |  11.27 GiB |    20.91 B | CUDA       |  99 |    tg100 @ d100 |        245.18 ± 0.66 |
| qwen3moe 30B.A3B Q3_K - Small  |  12.37 GiB |    30.53 B | CUDA       |  99 |    pp100 @ d100 |      1908.57 ± 27.90 |
| qwen3moe 30B.A3B Q3_K - Small  |  12.37 GiB |    30.53 B | CUDA       |  99 |    tg100 @ d100 |        170.90 ± 0.80 |