@ORippler 2025-09-02
> Thank you for the PR, this is potentially useful for other kernels as well.

Agreed, one kernel that came to my mind was [`k_bin_bcast`](https://github.com/ggml-org/llama.cpp/blob/9961d244f2df6baf40af2f1ddc0927f8d91578c8/ggml/src/ggml-cuda/binbcast.cu#L29-L68).