@JohannesGaessler 2025-09-03
Thank you, I'll merge this once the CI is done. The ggml matrix multiplication / FlashAttention kernels need to do some integer divisions to determine which data to work on, but register pressure is also a major concern. I'll investigate the use of `fastdiv` for those cases (unless you want to do it).