@JohannesGaessler 2025-09-03
For the cases relevant for GEMM/FA I think it will be possible to store `mp` with 16 bits, L with 4 bits, and the original value with 12 bits. That way you would still need only a single register to store the data.