@JohannesGaessler 2025-09-03
The benefit seems to be larger for fast GPUs:

| GPU      | Model                     |     Microbatch size | Test     |     t/s b6341 |     t/s 8bde72b51 |     Speedup |
| :------  | :------------------------ | ------------------: | :------- | ------------: | ----------------: | ----------: |
| P40      | llama 8B Q4_0             |                   1 | pp512    |         56.22 |             56.32 |        1.00 |
| P40      | llama 8B Q4_0             |                   2 | pp512    |        111.05 |            111.59 |        1.00 |
| P40      | llama 8B Q4_0             |                   4 | pp512    |        159.18 |            159.70 |        1.00 |
| P40      | llama 8B Q4_0             |                   8 | pp512    |        199.42 |            201.18 |        1.01 |
| P40      | llama 8B Q4_0             |                  16 | pp512    |        473.35 |            475.96 |        1.01 |
| P40      | llama 8B Q4_0             |                  32 | pp512    |        577.08 |            578.60 |        1.00 |
| P40      | llama 8B Q4_0             |                  64 | pp512    |        779.56 |            782.74 |        1.00 |
| P40      | llama 8B Q4_0             |                 128 | pp512    |        901.20 |            900.61 |        1.00 |
| P40      | llama 8B Q4_0             |                 256 | pp512    |        989.08 |            990.53 |        1.00 |
| P40      | llama 8B Q4_0             |                 512 | pp512    |       1033.35 |           1034.60 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                   1 | pp512    |         45.52 |             46.18 |        1.01 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                   2 | pp512    |         68.14 |             68.68 |        1.01 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                   4 | pp512    |        103.04 |            103.29 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                   8 | pp512    |        155.04 |            155.08 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                  16 | pp512    |        262.75 |            263.64 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                  32 | pp512    |        334.52 |            336.49 |        1.01 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                  64 | pp512    |        367.55 |            368.87 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                 128 | pp512    |        562.84 |            565.18 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                 256 | pp512    |        782.93 |            783.50 |        1.00 |
| P40      | qwen3moe 30B.A3B Q3_K_S   |                 512 | pp512    |       1003.04 |           1005.36 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                   1 | pp512    |        188.47 |            189.43 |        1.01 |
| RTX 4090 | llama 8B Q4_0             |                   2 | pp512    |        334.45 |            335.15 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                   4 | pp512    |        653.64 |            653.80 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                   8 | pp512    |       1085.25 |           1085.87 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                  16 | pp512    |       1842.49 |           1847.15 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                  32 | pp512    |       3332.12 |           3351.60 |        1.01 |
| RTX 4090 | llama 8B Q4_0             |                  64 | pp512    |       5814.07 |           5821.98 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                 128 | pp512    |       8733.03 |           8777.78 |        1.01 |
| RTX 4090 | llama 8B Q4_0             |                 256 | pp512    |      11730.38 |          11773.61 |        1.00 |
| RTX 4090 | llama 8B Q4_0             |                 512 | pp512    |      13258.03 |          13247.95 |        1.00 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                   1 | pp512    |        224.25 |            229.86 |        1.03 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                   2 | pp512    |        216.71 |            219.06 |        1.01 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                   4 | pp512    |        374.11 |            377.80 |        1.01 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                   8 | pp512    |        611.66 |            614.01 |        1.00 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                  16 | pp512    |        976.81 |            982.17 |        1.01 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                  32 | pp512    |       1557.01 |           1567.95 |        1.01 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                  64 | pp512    |       2546.91 |           2548.34 |        1.00 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                 128 | pp512    |       3464.85 |           3469.72 |        1.00 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                 256 | pp512    |       5594.30 |           5582.37 |        1.00 |
| RTX 4090 | qwen3moe 30B.A3B Q4_0     |                 512 | pp512    |       7856.42 |           7843.87 |        1.00 |