@ORippler 2025-09-04
> The benefit seems to be larger for fast GPUs:

Indeed, but only because for them latency-limited kernels such as `rms_norm_f32` take up more relative share of the overall workload.
<img width="2544" height="1329" alt="image" src="https://github.com/user-attachments/assets/94679efe-f4f9-48e7-a2b7-c530851bbcb3" />

> Thank you, I'll merge this once the CI is done. The ggml matrix multiplication / FlashAttention kernels need to do some integer divisions to determine which data to work on, but register pressure is also a major concern. I'll investigate the use of fastdiv for those cases (unless you want to do it).

Thanks for merging! I still think the `fastdiv` pattern is promising and should be applied to more kernels, as all GPUs will benefit from it perf-wise. I'll personally be available to work on this ~middle of October earliest as I'm going to take a leave, but will gladly assist in any way I can until my leave.

Regarding kernel pressure: you have to additionally factor in the registers required by the naive integer division to handle its intermediates. Playing around in [godbolt](https://godbolt.org/z/YTrzrfs8v) (so static analysis only) I couldn't find out a use-case where fastdiv used more registers for the SMs I compiled SASS for.