@JohannesGaessler 2025-09-04
Using `fastdiv` for `mul_mat_vec_q` and `quantize_q8_1` shaves off ~1% for token generation on RTX 3090/4090: https://github.com/ggml-org/llama.cpp/pull/15802